Renamed pages/dashboard to pages/search. Update all path references to /dashboard to /search.
