+++ frontend/tests/snapshots/search.test.tsx
@@ -1,5 +1,5 @@
 import * as React from "react"
-import Dashboard from "../../pages/dashboard"
+import Dashboard from "../../pages/search"
 import renderer from "react-test-renderer"
 
 it("renders Dashboard correctly", () => {
